Insert manifest within a transaction

+++ rhizome.h
@@ -243,7 +243,7 @@ int rhizome_server_close_http_request(int i);
 int rhizome_server_http_send_bytes(int rn,rhizome_http_request *r);
 int rhizome_server_parse_http_request(int rn,rhizome_http_request *r);
 int rhizome_server_simple_http_response(rhizome_http_request *r,int result, char *response);
-long long sqlite_exec_void(const char *sqlformat,...);
+int sqlite_exec_void(const char *sqlformat,...);
 long long sqlite_exec_int64(const char *sqlformat,...);
 int sqlite_exec_strbuf(strbuf sb, const char *sqlformat,...);
 int rhizome_server_http_response_header(rhizome_http_request *r,int result,
